Update PetCare.csproj
The <FileAlignment> element specifies the alignment of file positions in the output file, which can optimize file access and loading performance. Here, it's set to "512.". The <AutoGenerateBindingRedirects> element controls whether the build process automatically generates binding redirects in the application's configuration file. It's set to "true", indicating that binding redirects are automatically generated.The <Deterministic> element ensures deterministic builds, meaning that the output are identical given the same source code and build in environment.The <PropertyGroup> element with a condition specifying the 'Debug' configuration for 'AnyCPU' platform contains various settings specific to debugging. It defines the target platform as 'AnyCPU', enables debugging symbols and full debug information, disables optimization, specifies the output path for debug binaries, defines debug constants, sets error reporting to prompt, and sets warning level to 4.