Update PetCare.csproj
Startup Object line specifies which part of the application should be executed first when the program starts. In this case, it's set to 'PetCare.Program', indicating that the 'Main' method in the 'Program' class within the 'PetCare' namespace should be executed first. The application's icon file is specified as 'logoapp.ico'. This is the image that presents the application in the user interface, such as on the desktop or in the taskbar.
Reference section lists external libraries or components that the PETCare application depends on. In this case, it includes a reference to a library called 'Guna.UI2',.The 'HintPath' specifies the location of the library's DLL file on the developer's machine.