Update UserDash.cs
addProfilePicBtn_Click method is called when the button is clicked. It takes two parameters Object sender and event args.
OpenFileDialog is a built-in .NET class that allows users to browse and select files from their computer.
if (opendlg.ShowDialog() == DialogResult.OK) displays the file dialog created earlier and checks if the user clicked the "OK" button.